Dodano nowe funkcje walidacji i limitowania żądań: globalLimiter, authRateLimiter, sendLimiter, validateJsonPayload, sanitizeInput, fileTypeValidator.
